backend_server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
pymysql
google-genai
alembic
python-multipart
uvicorn
